[format] Fix clang-format-14 violations and CodeQL integer-overflow findings
Apply clang-format-14 to changed lines in optimizer_devel.h and
nntrainer-api-common.h to pass the cpp-linter CI check. In
Applications/MeZO/jni/main.cpp cast the loop index to size_t before
multiplying by the field-size members so the arithmetic is done in
size_t, preventing unsigned int overflow before implicit widening to
difference_type in iterator arithmetic (CodeQL rule
multiplication-result-converted-to-larger-type).

+++ Applications/MeZO/jni/main.cpp
@@ -74,12 +74,14 @@ int getSample(float **outVec, float **outLabel, bool *last, void *user_data) {
   unsigned int idx = ud->indices[ud->current_index];
 
   // Copy input data
-  std::copy(ud->data.begin() + idx * ud->feature_size,
-            ud->data.begin() + (idx + 1) * ud->feature_size, *outVec);
+  std::copy(ud->data.begin() + static_cast<size_t>(idx) * ud->feature_size,
+            ud->data.begin() + static_cast<size_t>(idx + 1) * ud->feature_size,
+            *outVec);
 
   // Copy label data
-  std::copy(ud->labels.begin() + idx * ud->label_size,
-            ud->labels.begin() + (idx + 1) * ud->label_size, *outLabel);
+  std::copy(ud->labels.begin() + static_cast<size_t>(idx) * ud->label_size,
+            ud->labels.begin() + static_cast<size_t>(idx + 1) * ud->label_size,
+            *outLabel);
 
   ud->current_index++;
